+++ core/view.js
@@ -150,7 +150,7 @@ View.prototype.setPosition = function (pos) {
         this.position.col = parseInt(arguments[1], 10);
     }
 
-    //  sanatize
+    //  sanitize
     this.position.row = Math.max(this.position.row, 1);
     this.position.col = Math.max(this.position.col, 1);
     this.position.row = Math.min(this.position.row, this.client.term.termHeight);